#0: Created longer nightly test for conv3d, shortened post-commit
